refactor(log-levels): extract reusable overflow-list and chip components
- add `<overflow-list>`: slot-based responsive overflow row that measures slotted children and moves the ones that dont fit into a native popover menu (Blueprint-style collapse-from/min-visible/menu-heading + ::part hooks)
- add `<vs-chip>`: presentational dropdown-face chip (lead slot + value)
- reduce LogLevels to a thin adapter rendering `<overflow-list>` of `<vs-chip>`, keeping the migration-to-filters seam
- rename logLevelsOverflow → overflowFit (helper is now generic)
